requirements.txt
PyQt6==6.6.1
python-docx==1.1.2
requests==2.32.3
